rename `smt_new`
